# textparse: Optimized protobuf parser with custom streaming unmarshal.