Fixes https://github.com/prometheus/prometheus/issues/14668

Depends on https://github.com/prometheus/prometheus/pull/15966

Previous attempts https://github.com/prometheus/prometheus/pull/15726 https://github.com/prometheus/prometheus/pull/15729

## Design

We have opportunity to stream protobuf unmarshalling. It's already sort of streamed by metric family, however no memory is reused across families. With this new implementation we:

1. Have a single `MetricSteamingDecoder` that allows progressing per MetricFamily.  It reuses previous MetricFamily object (e.g. internal bytes for lazy decoding of metrics).
2. It also allows progressing per Metric, by unmarshalling MetricFamilies WITHOUT metrics array. We only save bytes position of each metric to unmarshal it later.
3. When unmarshalling Metric we skip labels (we save position of each label in serialized proto for later). It offers separate Labels(*labels.ScratchBuilder) to parse and put labels directly into builder.
4. All strings are yoloStrings, no copying. The only "copy" is for labels as we use them later in appending, so they can't be reused.

## Maintenance

While it looks cumbersome to maintain, to make things efficient, one way or another we have to have streaming/lazy decoding with reusing of as much as possible. This does not depend on gogo we can do this on OpaqueAPI too but didn't want to change too much here. We could try to work with vtprotobuf or write our own plugin for those methods, but we can start by handcrafting now, this proto is not changing at all (if it will change in future, it will be OpenMetric proto likely).