@bwplotka 2024-12-30
I have to fix tests etc, but initial benchmarks looks promising. There might be more to optimize here, didn't look on this iteration profile yet.

```
benchstat -filter="/fmt:PromProto" gogo=scrape-loop-v1.txt custom=scrape-loop-customdecode.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                 │    gogo     │               custom               │
                                 │   sec/op    │   sec/op     vs base               │
ScrapeLoopAppend/fmt=PromProto-2   54.48µ ± 8%   23.43µ ± 9%  -56.99% (p=0.002 n=6)

                                 │     gogo     │               custom                │
                                 │     B/op     │     B/op      vs base               │
ScrapeLoopAppend/fmt=PromProto-2   75.58Ki ± 1%   26.88Ki ± 6%  -64.44% (p=0.002 n=6)

                                 │    gogo    │              custom               │
                                 │ allocs/op  │ allocs/op   vs base               │
ScrapeLoopAppend/fmt=PromProto-2   810.0 ± 0%   123.0 ± 0%  -84.81% (p=0.002 n=6)
```

Now proto does not look too bad:

```
 benchstat -col /fmt  scrape-loop-customdecode.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                   │  PromText   │              OMText               │             PromProto              │
                   │   sec/op    │   sec/op     vs base              │   sec/op     vs base               │
ScrapeLoopAppend-2   33.96µ ± 6%   30.88µ ± 3%  -9.08% (p=0.002 n=6)   23.43µ ± 9%  -31.01% (p=0.002 n=6)

                   │   PromText   │            OMText             │              PromProto              │
                   │     B/op     │     B/op      vs base         │     B/op      vs base               │
ScrapeLoopAppend-2   19.00Ki ± 7%   19.67Ki ± 6%  ~ (p=0.937 n=6)   26.88Ki ± 6%  +41.47% (p=0.002 n=6)

                   │  PromText  │              OMText              │              PromProto              │
                   │ allocs/op  │ allocs/op   vs base              │  allocs/op   vs base                │
ScrapeLoopAppend-2   12.00 ± 0%   13.00 ± 0%  +8.33% (p=0.002 n=6)   123.00 ± 0%  +925.00% (p=0.002 n=6)
```