@bwplotka 2025-02-03
This should be now good to go 🎉 

cc @krajorama @beorn7 @bboreham 

Results looks solid. There's room for further allocation improvements, but I propose we iterate. Some of those needs interface updates.

```
$ benchstat append-v1.txt append-v2.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                                     │ append-v1.txt │           append-v2.txt            │
                                                     │    sec/op     │   sec/op     vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromText-2     280.4µ ± 10%   283.9µ ± 8%        ~ (p=0.937 n=6)
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=OMText-2       265.1µ ± 10%   271.7µ ± 7%        ~ (p=0.818 n=6)
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    447.0µ ±  2%   299.8µ ± 7%  -32.93% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromText-2     168.9µ ±  3%   168.6µ ± 2%        ~ (p=0.937 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=OMText-2       166.4µ ±  1%   167.5µ ± 2%   +0.68% (p=0.041 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromProto-2    159.0µ ±  3%   146.1µ ± 2%   -8.07% (p=0.002 n=6)
geomean                                                 230.1µ         213.8µ        -7.10%

                                                     │ append-v1.txt │            append-v2.txt            │
                                                     │     B/op      │     B/op      vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromText-2    178.7Ki ±  8%   182.3Ki ± 6%        ~ (p=0.240 n=6)
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=OMText-2      170.5Ki ± 10%   172.3Ki ± 2%        ~ (p=0.937 n=6)
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2   665.7Ki ±  2%   212.5Ki ± 3%  -68.08% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromText-2    74.02Ki ±  1%   73.53Ki ± 3%        ~ (p=0.589 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=OMText-2      72.87Ki ±  1%   73.53Ki ± 3%        ~ (p=0.485 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromProto-2   169.2Ki ±  1%   110.6Ki ± 5%  -34.62% (p=0.002 n=6)
geomean                                                162.6Ki         125.9Ki       -22.56%

                                                     │ append-v1.txt │            append-v2.txt            │
                                                     │   allocs/op   │ allocs/op   vs base                 │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromText-2       11.00 ± 0%   11.00 ± 0%        ~ (p=1.000 n=6) ¹
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=OMText-2         12.00 ± 0%   12.00 ± 0%        ~ (p=1.000 n=6) ¹
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    8012.00 ± 0%   21.00 ± 0%  -99.74% (p=0.002 n=6)
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromText-2       18.00 ± 0%   18.00 ± 0%        ~ (p=1.000 n=6) ¹
ScrapeLoopAppend/data=59FamsAllTypes/fmt=OMText-2         19.00 ± 0%   19.00 ± 0%        ~ (p=1.000 n=6) ¹
ScrapeLoopAppend/data=59FamsAllTypes/fmt=PromProto-2     1693.0 ± 0%   642.0 ± 0%  -62.08% (p=0.002 n=6)
geomean                                                   92.15        29.11       -68.41%
¹ all samples are equal
```