@bwplotka 2025-02-04
I have some ideas to optimize further (see https://pprof.me/14c6981f2f1f0ab8dce893729df1da67/?profileType=profile%253Aalloc_objects%253Acount%253Aspace%253Abytes&color_by=filename), but I will do those in the next iteration - this PR is already big.