@bwplotka 2025-02-04
/prombench main --bench.version=bench/protofirst