@bwplotka 2025-02-05
Something got stuck (filled https://github.com/prometheus/test-infra/issues/831), retrying prombench

/prombench main --bench.version=bench/protofirst