@bwplotka 2025-02-05
eh what a mess with our [testdata](https://github.com/prometheus/prometheus/blob/8d281c3491564738aab1f6735eea4542428d999f/model/textparse/testdata/promtestdata.txt) - just noticed it does not have histograms. So my microbenchmark does not include those. I will need to create a separate benchmark improvement PR and rebase this one it. :/  Might need to clean up textparse/testdata first.