@bwplotka 2025-02-05
Prombench is dead at the moment https://github.com/prometheus/test-infra/issues/832