@bwplotka 2025-02-05
https://cloud-native.slack.com/archives/C01AUBA4PFE/p1738594684420579 @krajorama reported some problems when testing this PR, to be investigated

> I have 7000+ metrics, only ~300 reporting issues with duplicates timestamp different value