@bwplotka 2025-02-10
Rebased on top of https://github.com/prometheus/prometheus/pull/15998 for additional benchmark data