@bwplotka 2025-02-10
Latest benchmarks:

```
benchstat v1.txt v2.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/textparse
cpu: Apple M1 Pro
                 │   v1.txt    │               v2.txt               │
                 │   sec/op    │   sec/op     vs base               │
ParsePromProto-2   40.50µ ± 5%   28.89µ ± 2%  -28.66% (p=0.002 n=6)

                 │    v1.txt    │                v2.txt                │
                 │     B/s      │      B/s       vs base               │
ParsePromProto-2   90.55Mi ± 6%   126.92Mi ± 2%  +40.18% (p=0.002 n=6)

                 │    v1.txt    │               v2.txt                │
                 │     B/op     │     B/op      vs base               │
ParsePromProto-2   40.14Ki ± 0%   26.41Ki ± 0%  -34.20% (p=0.002 n=6)

                 │   v1.txt   │              v2.txt               │
                 │ allocs/op  │ allocs/op   vs base               │

```

```
benchstat -filter "/fmt:PromProto" append-v1.txt append-v2.txt
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                                      │ append-v1.txt │            append-v2.txt            │
                                                      │    sec/op     │    sec/op     vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2      957.1µ ± 8%   607.2µ ±  5%  -36.56% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2     871.5µ ± 6%   813.9µ ± 11%   -6.61% (p=0.041 n=6)
geomean                                                   913.3µ        703.0µ        -23.03%

                                                      │ append-v1.txt │            append-v2.txt            │
                                                      │     B/op      │     B/op      vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    1483.7Ki ± 1%   458.0Ki ± 6%  -69.13% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2    942.4Ki ± 0%   615.3Ki ± 3%  -34.71% (p=0.002 n=6)
geomean                                                  1.155Mi        530.9Ki       -55.11%

                                                      │ append-v1.txt │           append-v2.txt            │
                                                      │   allocs/op   │  allocs/op   vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    16017.00 ± 0%    25.00 ± 4%  -99.84% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2     8.168k ± 0%   3.047k ± 0%  -62.70% (p=0.002 n=6)
geomean                                                   11.44k         276.0       -97.59%
```

Compared to other parsers now:

```
 benchstat -col /fmt  append-v2.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                        │   PromText   │               OMText               │             PromProto              │
                                        │    sec/op    │    sec/op     vs base              │    sec/op     vs base              │
ScrapeLoopAppend/data=1Fam1000Gauges-2    594.5µ ± 29%   568.8µ ± 12%       ~ (p=0.485 n=6)   607.2µ ±  5%       ~ (p=0.699 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   815.8µ ±  6%   887.3µ ±  4%  +8.77% (p=0.009 n=6)   813.9µ ± 11%       ~ (p=1.000 n=6)
geomean                                   696.4µ         710.4µ        +2.01%                 703.0µ        +0.95%

                                        │   PromText    │               OMText                │              PromProto              │
                                        │     B/op      │     B/op       vs base              │     B/op      vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges-2    347.9Ki ±  3%   352.1Ki ±  8%       ~ (p=0.818 n=6)   458.0Ki ± 6%  +31.63% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   315.5Ki ± 12%   308.3Ki ± 10%       ~ (p=0.699 n=6)   615.3Ki ± 3%  +95.04% (p=0.002 n=6)
geomean                                   331.3Ki         329.5Ki        -0.54%                 530.9Ki       +60.23%

                                        │  PromText  │              OMText              │               PromProto                │
                                        │ allocs/op  │ allocs/op   vs base              │  allocs/op    vs base                  │
ScrapeLoopAppend/data=1Fam1000Gauges-2    13.00 ± 0%   13.50 ± 4%       ~ (p=0.182 n=6)     25.00 ± 4%     +92.31% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   14.00 ± 0%   15.00 ± 0%  +7.14% (p=0.002 n=6)   3047.00 ± 0%  +21664.29% (p=0.002 n=6)
geomean                                   13.49        14.23       +5.48%                   276.0        +1945.83%

```