@bwplotka 2025-02-10
Ok, after reusing complex types and help bytes, I think it's good for this PR. We can optimize more later on, majority of this are labels.Labels and magic suffixes which might require protocol/parser interface redesign.

![image](https://github.com/user-attachments/assets/bb12e466-7c01-4e77-a1e3-8b54ba2fea42)

Final results:

```
 benchstat v1.txt v4.txt                           
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/textparse
cpu: Apple M1 Pro
                 │   v1.txt    │               v4.txt               │
                 │   sec/op    │   sec/op     vs base               │
ParsePromProto-2   40.50µ ± 5%   26.22µ ± 3%  -35.26% (p=0.002 n=6)

                 │    v1.txt    │                v4.txt                │
                 │     B/s      │      B/s       vs base               │
ParsePromProto-2   90.55Mi ± 6%   139.87Mi ± 3%  +54.47% (p=0.002 n=6)

                 │    v1.txt    │               v4.txt                │
                 │     B/op     │     B/op      vs base               │
ParsePromProto-2   40.14Ki ± 0%   20.82Ki ± 0%  -48.12% (p=0.002 n=6)

                 │   v1.txt   │              v4.txt               │
                 │ allocs/op  │ allocs/op   vs base               │
ParsePromProto-2   870.0 ± 0%   465.0 ± 0%  -46.55% (p=0.002 n=6)
```

```
benchstat -filter "/fmt:PromProto" append-v1.txt append-v4.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                                      │ append-v1.txt │            append-v4.txt            │
                                                      │    sec/op     │    sec/op     vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2      957.1µ ± 8%   572.6µ ± 10%  -40.17% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2     871.5µ ± 6%   733.0µ ±  4%  -15.89% (p=0.002 n=6)
geomean                                                   913.3µ        647.9µ        -29.06%

                                                      │ append-v1.txt │            append-v4.txt            │
                                                      │     B/op      │     B/op      vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    1483.7Ki ± 1%   461.4Ki ± 5%  -68.90% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2    942.4Ki ± 0%   397.2Ki ± 4%  -57.85% (p=0.002 n=6)
geomean                                                  1.155Mi        428.1Ki       -63.80%

                                                      │ append-v1.txt │           append-v4.txt            │
                                                      │   allocs/op   │  allocs/op   vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges/fmt=PromProto-2    16017.00 ± 0%    23.50 ± 2%  -99.85% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes/fmt=PromProto-2     8.168k ± 0%   2.226k ± 0%  -72.75% (p=0.002 n=6)
geomean                                                   11.44k         228.7       -98.00%

```

Compared to other parsers now:
```
benchstat -col /fmt  append-v4.txt 
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/scrape
cpu: Apple M1 Pro
                                        │   PromText   │               OMText               │             PromProto              │
                                        │    sec/op    │    sec/op     vs base              │    sec/op     vs base              │
ScrapeLoopAppend/data=1Fam1000Gauges-2    613.0µ ± 28%   585.2µ ± 10%  -4.54% (p=0.041 n=6)   572.6µ ± 10%  -6.59% (p=0.041 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   780.1µ ±  4%   780.3µ ±  8%       ~ (p=0.818 n=6)   733.0µ ±  4%  -6.04% (p=0.009 n=6)
geomean                                   691.5µ         675.7µ        -2.29%                 647.9µ        -6.31%

                                        │   PromText    │               OMText                │              PromProto              │
                                        │     B/op      │     B/op       vs base              │     B/op      vs base               │
ScrapeLoopAppend/data=1Fam1000Gauges-2    341.4Ki ±  7%   354.8Ki ±  5%       ~ (p=0.394 n=6)   461.4Ki ± 5%  +35.15% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   311.9Ki ± 12%   319.2Ki ± 11%       ~ (p=0.485 n=6)   397.2Ki ± 4%  +27.36% (p=0.002 n=6)
geomean                                   326.3Ki         336.5Ki        +3.13%                 428.1Ki       +31.19%

                                        │  PromText  │              OMText              │               PromProto                │
                                        │ allocs/op  │ allocs/op   vs base              │  allocs/op    vs base                  │
ScrapeLoopAppend/data=1Fam1000Gauges-2    13.00 ± 8%   13.00 ± 8%       ~ (p=0.273 n=6)     23.50 ± 2%     +80.77% (p=0.002 n=6)
ScrapeLoopAppend/data=237FamsAllTypes-2   14.00 ± 0%   15.00 ± 0%  +7.14% (p=0.002 n=6)   2226.00 ± 0%  +15800.00% (p=0.002 n=6)
geomean                                   13.49        13.96       +3.51%                   228.7        +1595.36%

```